requirements.txt
anyio==3.7.0
backoff==2.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
chromadb==0.3.25
click==8.1.3
clickhouse-connect==0.5.25
coloredlogs==15.0.1
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.35
duckdb==0.8.0
fastapi==0.95.2
flatbuffers==23.5.26
fonttools==4.39.4
h11==0.14.0
hdbscan==0.8.29
hnswlib==0.7.0
httptools==0.5.0
humanfriendly==10.0
idna==3.4
joblib==1.2.0
kiwisolver==1.4.4
llvmlite==0.40.0
lz4==4.3.2
matplotlib==3.7.1
monotonic==1.6
mpmath==1.3.0
numba==0.57.0
numpy==1.24.3
onnxruntime==1.15.0
overrides==7.3.1
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
posthog==3.0.1
protobuf==4.23.2
pydantic==1.10.8
pynndescent==0.5.10
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
threadpoolctl==3.1.0
tokenizers==0.13.3
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
umap-learn==0.5.3
urllib3==2.0.2
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
zstandard==0.21.0
